SF:Report active config to java framework
When config multi display config in display driver, the active
config in java framework is not same as SF. Without the Changed
flag the active config can not report to java framework.During
the booting, will set active config and report to java framework,
when boot finished, it may change the active config, so we should
add the Changed flag, and report java framework.

Test: make, take a photo, record a video and play video,
	run vts testRefreshRate

adyabr@ Note: This fix is not needed on master as we are not
changing the refresh rate in bootFinished. Added Merged-In to reference
the change on master.

Change-Id: I2e48640f48a14bc044884c2e834af53e9a87ef13
Merged-In: I5bb964572d93e5cb78d7b75054b900917be563b0
Signed-off-by: luochaojiang <luochaojiang@xiaomi.com>
Signed-off-by: AnkitGourav <Ankitgourav035@gmail.com>